requirements.txt
fastapi[standard]
kafka-python-ng
numpy
sse-starlette